fix: poll lp-token balances only in yield widget
