+++ apps/cowswap-frontend/src/common/updaters/LpBalancesAndAllowancesUpdater.tsx
@@ -1,0 +1,51 @@
+import { useEffect, useMemo, useState } from 'react'
+
+import { usePersistBalancesAndAllowances } from '@cowprotocol/balances-and-allowances'
+import { SWR_NO_REFRESH_OPTIONS } from '@cowprotocol/common-const'
+import type { SupportedChainId } from '@cowprotocol/cow-sdk'
+import { TokenListCategory, useAllLpTokens } from '@cowprotocol/tokens'
+
+import ms from 'ms.macro'
+
+// A small gap between balances and allowances refresh intervals is needed to avoid high load to the node at the same time
+const LP_BALANCES_SWR_CONFIG = { refreshInterval: ms`32s` }
+const LP_ALLOWANCES_SWR_CONFIG = { refreshInterval: ms`34s` }
+const LP_MULTICALL_OPTIONS = { consequentExecution: true }
+
+// To avoid high load to the node at the same time
+// We start the updater with a delay
+const LP_UPDATER_START_DELAY = ms`3s`
+
+const LP_CATEGORIES = [TokenListCategory.LP, TokenListCategory.COW_AMM_LP]
+
+export interface BalancesAndAllowancesUpdaterProps {
+  account: string | undefined
+  chainId: SupportedChainId
+  enablePolling: boolean
+}
+export function LpBalancesAndAllowancesUpdater({ account, chainId, enablePolling }: BalancesAndAllowancesUpdaterProps) {
+  const allLpTokens = useAllLpTokens(LP_CATEGORIES)
+  const [isUpdaterPaused, setIsUpdaterPaused] = useState(true)
+
+  const lpTokenAddresses = useMemo(() => allLpTokens.map((token) => token.address), [allLpTokens])
+
+  usePersistBalancesAndAllowances({
+    account: isUpdaterPaused ? undefined : account,
+    chainId,
+    tokenAddresses: lpTokenAddresses,
+    setLoadingState: false,
+    balancesSwrConfig: enablePolling ? LP_BALANCES_SWR_CONFIG : SWR_NO_REFRESH_OPTIONS,
+    allowancesSwrConfig: enablePolling ? LP_ALLOWANCES_SWR_CONFIG : SWR_NO_REFRESH_OPTIONS,
+    multicallOptions: LP_MULTICALL_OPTIONS,
+  })
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setIsUpdaterPaused(false)
+    }, LP_UPDATER_START_DELAY)
+
+    return () => clearTimeout(timeout)
+  }, [])
+
+  return null
+}
