+++ apps/cowswap-frontend/src/modules/application/containers/App/Updaters.tsx
@@ -13,6 +13,7 @@ import { FinalizeTxUpdater } from 'modules/onchainTransactions'
 import { OrdersNotificationsUpdater } from 'modules/orders'
 import { EthFlowDeadlineUpdater } from 'modules/swap/state/EthFlow/updaters'
 import { useOnTokenListAddingError } from 'modules/tokensList'
+import { TradeType, useTradeTypeInfo } from 'modules/trade'
 import { UsdPricesUpdater } from 'modules/usdAmount'
 
 import { ProgressBarV2ExecutingOrdersUpdater } from 'common/hooks/orderProgressBarV2'
@@ -20,6 +21,7 @@ import { TotalSurplusUpdater } from 'common/state/totalSurplusState'
 import { FeatureFlagsUpdater } from 'common/updaters/FeatureFlagsUpdater'
 import { FeesUpdater } from 'common/updaters/FeesUpdater'
 import { GasUpdater } from 'common/updaters/GasUpdater'
+import { LpBalancesAndAllowancesUpdater } from 'common/updaters/LpBalancesAndAllowancesUpdater'
 import {
   CancelledOrdersUpdater,
   ExpiredOrdersUpdater,
@@ -36,6 +38,8 @@ export function Updaters() {
   const { tokenLists, appCode, customTokens, standaloneMode } = useInjectedWidgetParams()
   const onTokenListAddingError = useOnTokenListAddingError()
   const { isGeoBlockEnabled } = useFeatureFlags()
+  const tradeTypeInfo = useTradeTypeInfo()
+  const isYieldWidget = tradeTypeInfo?.tradeType === TradeType.YIELD
 
   return (
     <>
@@ -76,6 +80,7 @@ export function Updaters() {
       />
       <UnsupportedTokensUpdater />
       <BalancesAndAllowancesUpdater chainId={chainId} account={account} />
+      <LpBalancesAndAllowancesUpdater chainId={chainId} account={account} enablePolling={isYieldWidget} />
     </>
   )
 }
